requirements.txt
attrs==20.3.0
aws-cdk.assets==1.77.0
aws-cdk.aws-apigateway==1.77.0
aws-cdk.aws-apigatewayv2==1.77.0
aws-cdk.aws-applicationautoscaling==1.77.0
aws-cdk.aws-autoscaling==1.77.0
aws-cdk.aws-autoscaling-common==1.77.0
aws-cdk.aws-autoscaling-hooktargets==1.77.0
aws-cdk.aws-batch==1.77.0
aws-cdk.aws-certificatemanager==1.77.0
aws-cdk.aws-cloudformation==1.77.0
aws-cdk.aws-cloudfront==1.77.0
aws-cdk.aws-cloudwatch==1.77.0
aws-cdk.aws-codebuild==1.77.0
aws-cdk.aws-codecommit==1.77.0
aws-cdk.aws-codedeploy==1.77.0
aws-cdk.aws-codeguruprofiler==1.77.0
aws-cdk.aws-codepipeline==1.77.0
aws-cdk.aws-codepipeline-actions==1.77.0
aws-cdk.aws-cognito==1.77.0
aws-cdk.aws-ec2==1.77.0
aws-cdk.aws-ecr==1.77.0
aws-cdk.aws-ecr-assets==1.77.0
aws-cdk.aws-ecs==1.77.0
aws-cdk.aws-efs==1.77.0
aws-cdk.aws-elasticloadbalancing==1.77.0
aws-cdk.aws-elasticloadbalancingv2==1.77.0
aws-cdk.aws-events==1.77.0
aws-cdk.aws-events-targets==1.77.0
aws-cdk.aws-iam==1.77.0
aws-cdk.aws-kinesis==1.77.0
aws-cdk.aws-kinesisfirehose==1.77.0
aws-cdk.aws-kms==1.77.0
aws-cdk.aws-lambda==1.77.0
aws-cdk.aws-logs==1.77.0
aws-cdk.aws-route53==1.77.0
aws-cdk.aws-route53-targets==1.77.0
aws-cdk.aws-s3==1.77.0
aws-cdk.aws-s3-assets==1.77.0
aws-cdk.aws-sam==1.77.0
aws-cdk.aws-secretsmanager==1.77.0
aws-cdk.aws-servicecatalog==1.77.0
aws-cdk.aws-servicediscovery==1.77.0
aws-cdk.aws-sns==1.77.0
aws-cdk.aws-sns-subscriptions==1.77.0
aws-cdk.aws-sqs==1.77.0
aws-cdk.aws-ssm==1.77.0
aws-cdk.aws-stepfunctions==1.77.0
aws-cdk.cloud-assembly-schema==1.77.0
aws-cdk.core==1.77.0
aws-cdk.custom-resources==1.77.0
aws-cdk.cx-api==1.77.0
aws-cdk.pipelines==1.77.0
aws-cdk.region-info==1.77.0
cattrs==1.0.0
certifi==2020.12.5
chardet==3.0.4
constructs==3.2.65
idna==2.10
importlib-metadata==3.1.1
importlib-resources==3.3.0
iniconfig==1.1.1
jsii==1.16.0
packaging==20.7
pluggy==0.13.1
publication==0.0.3
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
zipp==3.4.0
